actualizacion de jugador

+++ BackEnd/src/routes/jugadorRuta.js
@@ -11,26 +11,26 @@ ruta.get('/jugador/:id', verifyToken, async(req, res)=>{
 });
 ruta.get('/jugadores/:id',verifyToken, async(req, res)=>{
     const {id}=req.params
-   const buscarEquipo= await Jugador.findById(id)
-   console.info(buscarEquipo)
-    res.send(buscarEquipo)
+   const buscarJugador= await Jugador.findById(id)
+    res.send(buscarJugador)
 });
-ruta.put('/jugador/:id',verifyToken, async(req, res)=>{
-    const {id} =req.params;
-    const jugador= req.body
+ruta.put('/jugador/:_id', async(req, res)=>{
+    const {_id} =req.params;
+    const {nombre, apellido}= req.body;
+    const newJugador= {nombre, apellido}  
     try {
-        const actualizarJugador= await Jugador.findByIdAndUpdate({_id: id }, jugador, {new :true}) 
+        const actualizarJugador= await Jugador.findByIdAndUpdate({_id}, newJugador, {new :true}) 
         if (actualizarJugador) {
             res.json({mesage: 'Jugador Actualizado'})
         }
     } catch (error) {
         console.log(error)
-    }   
+    } 
 })
-ruta.post('/jugador/:id',verifyToken, async(req, res)=>{ 
+ruta.post('/jugador/:id', async(req, res)=>{ 
+    console.log('aca llego jugador')
     const {id}= req.params
     const {nombre, apellido}= req.body
-    console.log(req.body)
     const jugador= new Jugador({nombre, apellido}) 
     const guardarJugador= await jugador.save()
     if (guardarJugador) { 
@@ -46,19 +46,25 @@ ruta.post('/jugador/:id',verifyToken, async(req, res)=>{
     }  }
     
 });
-ruta.delete('/jugador/:id',verifyToken, async(req, res)=>{
+ruta.delete('/jugador/:id', async(req, res)=>{
     const {id}= req.params
     const buscarEquipo= await Equipos.find( {jugadores: id} )
     if (buscarEquipo.length > 0) {
         const idEquipo=buscarEquipo[0]._id
-        await Equipos.updateOne({ _id: idEquipo}, {$pull: {jugadores: id}})
-           try {
-            await Jugador.findByIdAndRemove(id)
-            res.json({mesagge: 'Jugador Eliminado'})
-        } catch (error) {
-            console.log(error)   
+        const quitarJugador= await Equipos.updateOne({ _id: idEquipo}, {$pull: {jugadores: id}})
+        if (quitarJugador){
+            try {
+                await Jugador.findByIdAndRemove(id)
+                res.json({mesagge: 'Jugador Eliminado'})
+            } catch (error) {
+                console.log(error)  
+                res.status(401) 
+                return
+            }
         }
-    } res.send('No se encontro el jugador')
+           
+    }else{res.send('No se encontro el jugador');
+     return}
 })
 
 module.exports = ruta
